+++ cmd/cody-gateway/internal/httpapi/handler.go
@@ -25,18 +25,19 @@ import (
 )
 
 type Config struct {
-	RateLimitNotifier               notify.RateLimitNotifier
-	AnthropicAccessToken            string
-	AnthropicAllowedModels          []string
-	AnthropicAllowedPromptPatterns  []string
-	AnthropicRequestBlockingEnabled bool
-	AnthropicMaxTokensToSample      int
-	OpenAIAccessToken               string
-	OpenAIOrgID                     string
-	OpenAIAllowedModels             []string
-	FireworksAccessToken            string
-	FireworksAllowedModels          []string
-	EmbeddingsAllowedModels         []string
+	RateLimitNotifier                           notify.RateLimitNotifier
+	AnthropicAccessToken                        string
+	AnthropicAllowedModels                      []string
+	AnthropicAllowedPromptPatterns              []string
+	AnthropicRequestBlockingEnabled             bool
+	AnthropicMaxTokensToSample                  int
+	OpenAIAccessToken                           string
+	OpenAIOrgID                                 string
+	OpenAIAllowedModels                         []string
+	FireworksAccessToken                        string
+	FireworksAllowedModels                      []string
+	FireworksLogSelfServeCodeCompletionRequests bool
+	EmbeddingsAllowedModels                     []string
 }
 
 var meter = otel.GetMeterProvider().Meter("cody-gateway/internal/httpapi")
@@ -189,6 +190,7 @@ func NewHandler(
 								httpClient,
 								config.FireworksAccessToken,
 								config.FireworksAllowedModels,
+								config.FireworksLogSelfServeCodeCompletionRequests,
 							),
 						),
 					),