+++ cmd/frontend/internal/auth/bitbucketcloudoauth/config_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
+	"github.com/google/go-cmp/cmp/cmpopts"
 	"golang.org/x/oauth2"
 
 	"github.com/sourcegraph/log/logtest"
@@ -210,7 +211,7 @@ func TestParseConfig(t *testing.T) {
 			if diff := cmp.Diff(tt.wantProblems, gotProblems.Messages()); diff != "" {
 				t.Errorf("problems: %s", diff)
 			}
-			if diff := cmp.Diff(wantConfigs, gotConfigs); diff != "" {
+			if diff := cmp.Diff(wantConfigs, gotConfigs, cmpopts.IgnoreUnexported(oauth2.Config{})); diff != "" {
 				t.Errorf("problems: %s", diff)
 			}
 		})