+++ cmd/worker/internal/embeddings/repo/scheduler.go
@@ -83,6 +83,6 @@ func newRepoEmbeddingScheduler(
 		enqueueActive,
 		goroutine.WithName("repoEmbeddingSchedulerJob"),
 		goroutine.WithDescription("resolves embedding policies and schedules jobs to embed repos"),
-		goroutine.WithInterval(1*time.Minute),
+		goroutine.WithInterval(5*time.Minute),
 	)
 }